A zoom container which isn't a viewport (#121)
This commit adds components::ZoomContainer which is an object
which collaborates with a component that can internalize its draws
based on scale and starting position. Defacto this will let us
make the shortcircuit sample zoom way faster because we can just know
start points and end points.

The mouse gestures are, I think right. For a mac trackpad they
are glorious. But we will need to adjust them for win/lin mice
I'm sure.